+++ test/support/web/controllers/page_controller.ex
@@ -2,7 +2,7 @@ defmodule DrabTestApp.PageController do
   @moduledoc false
 
   use DrabTestApp.Web, :controller
-  use Drab.Controller
+  # use Drab.Controller
 
   def index(conn, _params) do
     render(conn, "index.html")